requirements.txt
dill
xgboost
Flask>=1.0
scikit-learn
pandas
gunicorn
gcloud==0.18.1
google-cloud-storage